Refactored *Configs into separate crates to resolve dependency cycles for #2007
