+++ chain/chain/Cargo.toml
@@ -19,6 +19,7 @@ lazy_static = "1.4"
 
 borsh = "0.2.10"
 
+near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
