+++ chain/client/src/lib.rs
@@ -4,7 +4,7 @@ extern crate lazy_static;
 pub use crate::client::Client;
 pub use crate::client_actor::ClientActor;
 pub use crate::types::{
-    ClientConfig, Error, GetBlock, GetChunk, GetGasPrice, GetKeyValueChanges, GetNetworkInfo,
+    Error, GetBlock, GetChunk, GetGasPrice, GetKeyValueChanges, GetNetworkInfo,
     GetNextLightClientBlock, GetValidatorInfo, Query, Status, StatusResponse, SyncStatus, TxStatus,
 };
 pub use crate::view_client::ViewClientActor;
