+++ chain/network/Cargo.toml
@@ -23,6 +23,7 @@ tracing = "0.1"
 borsh = "0.2.10"
 cached = "0.11.0"
 
+near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
