+++ core/chain-configs/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "near-chain-configs"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[dependencies]
+chrono = { version = "0.4.4", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+near-crypto = { path = "../crypto" }
+near-primitives = { path = "../primitives" }
+near-runtime-configs = { path = "../runtime-configs" }
