+++ core/primitives/src/lib.rs
@@ -26,6 +26,7 @@ pub mod receipt;
 pub mod rpc;
 pub mod serialize;
 pub mod sharding;
+pub mod state_record;
 pub mod telemetry;
 pub mod test_utils;
 pub mod transaction;
