+++ core/runtime-configs/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "near-runtime-configs"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+
+near-primitives = { path = "../primitives" }
+near-runtime-fees = { path = "../../runtime/near-runtime-fees" }
+near-vm-logic = { path = "../../runtime/near-vm-logic" }
