+++ core/runtime-configs/src/lib.rs
@@ -1,0 +1,3 @@
+mod runtime_config;
+
+pub use runtime_config::RuntimeConfig;
