+++ genesis-tools/genesis-csv-to-json/src/csv_parser.rs
@@ -7,17 +7,16 @@ use chrono::{DateTime, Utc};
 use csv::ReaderBuilder;
 use serde::{Deserialize, Serialize};
 
-use near::config::AccountInfo;
 use near_crypto::{KeyType, PublicKey};
 use near_network::PeerInfo;
 use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::receipt::{ActionReceipt, Receipt, ReceiptEnum};
 use near_primitives::serialize::to_base64;
+use near_primitives::state_record::StateRecord;
 use near_primitives::transaction::{Action, FunctionCallAction};
-use near_primitives::types::{AccountId, Balance, Gas};
+use near_primitives::types::{AccountId, AccountInfo, Balance, Gas};
 use near_primitives::utils::{create_nonce_with_nonce, is_valid_account_id};
 use near_primitives::views::{AccessKeyPermissionView, AccessKeyView, AccountView};
-use node_runtime::StateRecord;
 
 /// Methods that can be called by a non-privileged access key.
 const REGULAR_METHOD_NAMES: &[&str] = &["stake", "transfer"];
