+++ genesis-tools/genesis-csv-to-json/src/csv_to_json_configs.rs
@@ -2,14 +2,13 @@ use std::fs::File;
 use std::path::Path;
 
 use near::config::{
-    get_initial_supply, Config, BLOCK_PRODUCER_KICKOUT_THRESHOLD, CHUNK_PRODUCER_KICKOUT_THRESHOLD,
-    CONFIG_FILENAME, CONFIG_VERSION, DEVELOPER_PERCENT, EXPECTED_EPOCH_LENGTH, FISHERMEN_THRESHOLD,
-    GAS_PRICE_ADJUSTMENT_RATE, GENESIS_CONFIG_FILENAME, INITIAL_GAS_LIMIT, MAX_INFLATION_RATE,
-    MIN_GAS_PRICE, NODE_KEY_FILE, NUM_BLOCKS_PER_YEAR, NUM_BLOCK_PRODUCER_SEATS, PROTOCOL_PERCENT,
-    TRANSACTION_VALIDITY_PERIOD,
+    Config, BLOCK_PRODUCER_KICKOUT_THRESHOLD, CHUNK_PRODUCER_KICKOUT_THRESHOLD, CONFIG_FILENAME,
+    DEVELOPER_PERCENT, EXPECTED_EPOCH_LENGTH, FISHERMEN_THRESHOLD, GAS_PRICE_ADJUSTMENT_RATE,
+    GENESIS_CONFIG_FILENAME, INITIAL_GAS_LIMIT, MAX_INFLATION_RATE, MIN_GAS_PRICE, NODE_KEY_FILE,
+    NUM_BLOCKS_PER_YEAR, NUM_BLOCK_PRODUCER_SEATS, PROTOCOL_PERCENT, TRANSACTION_VALIDITY_PERIOD,
 };
-use near::{GenesisConfig, NEAR_BASE};
-use near_network::types::PROTOCOL_VERSION;
+use near::NEAR_BASE;
+use near_chain_configs::{GenesisConfig, GENESIS_CONFIG_VERSION, PROTOCOL_VERSION};
 use near_primitives::types::{Balance, NumShards, ShardId};
 use near_primitives::utils::get_num_seats_per_shard;
 
@@ -54,19 +53,17 @@ pub fn csv_to_json_configs(home: &Path, chain_id: String, tracked_shards: Vec<Sh
         .expect("Error parsing accounts file.");
     config.network.boot_nodes =
         peer_info.into_iter().map(|x| x.to_string()).collect::<Vec<_>>().join(",");
-    let total_supply = get_initial_supply(&records);
-    verify_total_supply(total_supply, &chain_id);
-    let genesis_config = GenesisConfig {
+    let mut genesis_config = GenesisConfig {
         protocol_version: PROTOCOL_VERSION,
-        config_version: CONFIG_VERSION,
+        config_version: GENESIS_CONFIG_VERSION,
         genesis_time,
-        chain_id,
+        chain_id: chain_id.clone(),
         num_block_producer_seats: NUM_BLOCK_PRODUCER_SEATS,
         num_block_producer_seats_per_shard: get_num_seats_per_shard(
             NUM_SHARDS,
             NUM_BLOCK_PRODUCER_SEATS,
         ),
-        avg_hidden_validator_seats_per_shard: (0..NUM_SHARDS).map(|_| 0).collect(),
+        avg_hidden_validator_seats_per_shard: vec![0; NUM_SHARDS as usize],
         dynamic_resharding: false,
         epoch_length: EXPECTED_EPOCH_LENGTH,
         gas_limit: INITIAL_GAS_LIMIT,
@@ -79,13 +76,15 @@ pub fn csv_to_json_configs(home: &Path, chain_id: String, tracked_shards: Vec<Sh
         developer_reward_percentage: DEVELOPER_PERCENT,
         protocol_reward_percentage: PROTOCOL_PERCENT,
         max_inflation_rate: MAX_INFLATION_RATE,
-        total_supply,
+        total_supply: 0,
         num_blocks_per_year: NUM_BLOCKS_PER_YEAR,
         protocol_treasury_account: treasury,
         chunk_producer_kickout_threshold: CHUNK_PRODUCER_KICKOUT_THRESHOLD,
         min_gas_price: MIN_GAS_PRICE,
         fishermen_threshold: FISHERMEN_THRESHOLD,
     };
+    genesis_config.init();
+    verify_total_supply(genesis_config.total_supply, &chain_id);
 
     // Write all configs to files.
     config.write_to_file(&home.join(CONFIG_FILENAME));
