+++ genesis-tools/genesis-populate/Cargo.toml
@@ -12,10 +12,10 @@ clap = "2.33.0"
 tempdir = "0.3"
 
 near = { path = "../../near" }
+near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
-node-runtime = { path = "../../runtime/runtime" }
 near-chain = { path = "../../chain/chain" }
 near-client = { path = "../../chain/client" }
 near-pool = { path = "../../chain/pool" }
