+++ near/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 actix = "0.9.0"
 byteorder = "1.2"
+easy-ext = "0.1"
 rocksdb = "0.13"
 log = "0.4"
 env_logger = "0.7"
@@ -26,6 +27,8 @@ tracing-subscriber = "0.1.6"
 
 near-crypto = { path = "../core/crypto" }
 near-primitives = { path = "../core/primitives" }
+near-chain-configs = { path = "../core/chain-configs" }
+near-runtime-configs = { path = "../core/runtime-configs" }
 near-store = { path = "../core/store" }
 node-runtime = { path = "../runtime/runtime" }
 near-chain = { path = "../chain/chain" }
