+++ near/src/config.rs
@@ -1,36 +1,35 @@
 use std::convert::TryInto;
+use std::fs;
 use std::fs::File;
 use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
 use std::str;
 use std::sync::Arc;
 use std::time::Duration;
-use std::{cmp, fs};
 
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 use log::info;
 use serde_derive::{Deserialize, Serialize};
 
-use near_chain::ChainGenesis;
-use near_client::ClientConfig;
+use near_chain_configs::ClientConfig;
+use near_chain_configs::{GenesisConfig, GENESIS_CONFIG_VERSION, PROTOCOL_VERSION};
 use near_crypto::{InMemorySigner, KeyFile, KeyType, PublicKey, Signer};
 use near_jsonrpc::RpcConfig;
 use near_network::test_utils::open_port;
-use near_network::types::{PROTOCOL_VERSION, ROUTED_MESSAGE_TTL};
+use near_network::types::ROUTED_MESSAGE_TTL;
 use near_network::utils::blacklist_from_vec;
 use near_network::NetworkConfig;
 use near_primitives::account::AccessKey;
-use near_primitives::hash::{hash, CryptoHash};
-use near_primitives::serialize::{to_base64, u128_dec_format};
+use near_primitives::hash::CryptoHash;
+use near_primitives::state_record::StateRecord;
 use near_primitives::types::{
-    AccountId, Balance, BlockHeightDelta, Gas, NumBlocks, NumSeats, NumShards, ShardId,
+    AccountId, AccountInfo, Balance, BlockHeightDelta, Gas, NumBlocks, NumSeats, NumShards, ShardId,
 };
 use near_primitives::utils::{generate_random_string, get_num_seats_per_shard};
 use near_primitives::validator_signer::{InMemoryValidatorSigner, ValidatorSigner};
 use near_primitives::views::AccountView;
+use near_runtime_configs::RuntimeConfig;
 use near_telemetry::TelemetryConfig;
-use node_runtime::config::RuntimeConfig;
-use node_runtime::StateRecord;
 
 /// Initial balance used in tests.
 pub const TESTING_INIT_BALANCE: Balance = 1_000_000_000 * NEAR_BASE;
@@ -286,6 +285,89 @@ impl From<&str> for Config {
     }
 }
 
+#[easy_ext::ext(GenesisConfigExt)]
+impl GenesisConfig {
+    pub fn test_with_seeds(
+        seeds: Vec<&str>,
+        num_validator_seats: NumSeats,
+        num_validator_seats_per_shard: Vec<NumSeats>,
+    ) -> Self {
+        let mut validators = vec![];
+        let mut records = vec![];
+        for (i, &account) in seeds.iter().enumerate() {
+            let signer = InMemorySigner::from_seed(account, KeyType::ED25519, account);
+            let i = i as u64;
+            if i < num_validator_seats {
+                validators.push(AccountInfo {
+                    account_id: account.to_string(),
+                    public_key: signer.public_key.clone(),
+                    amount: TESTING_INIT_STAKE,
+                });
+            }
+            records.extend(
+                state_records_account_with_key(
+                    account,
+                    &signer.public_key.clone(),
+                    TESTING_INIT_BALANCE
+                        - if i < num_validator_seats { TESTING_INIT_STAKE } else { 0 },
+                    if i < num_validator_seats { TESTING_INIT_STAKE } else { 0 },
+                    CryptoHash::default(),
+                )
+                .into_iter(),
+            );
+        }
+        add_protocol_account(&mut records);
+        let mut config = GenesisConfig {
+            protocol_version: PROTOCOL_VERSION,
+            config_version: GENESIS_CONFIG_VERSION,
+            genesis_time: Utc::now(),
+            chain_id: random_chain_id(),
+            num_block_producer_seats: num_validator_seats,
+            num_block_producer_seats_per_shard: num_validator_seats_per_shard.clone(),
+            avg_hidden_validator_seats_per_shard: vec![0; num_validator_seats_per_shard.len()],
+            dynamic_resharding: false,
+            epoch_length: FAST_EPOCH_LENGTH,
+            gas_limit: INITIAL_GAS_LIMIT,
+            gas_price_adjustment_rate: GAS_PRICE_ADJUSTMENT_RATE,
+            block_producer_kickout_threshold: BLOCK_PRODUCER_KICKOUT_THRESHOLD,
+            runtime_config: Default::default(),
+            validators,
+            records,
+            developer_reward_percentage: DEVELOPER_PERCENT,
+            protocol_reward_percentage: PROTOCOL_PERCENT,
+            max_inflation_rate: MAX_INFLATION_RATE,
+            total_supply: 0,
+            num_blocks_per_year: NUM_BLOCKS_PER_YEAR,
+            protocol_treasury_account: PROTOCOL_TREASURY_ACCOUNT.to_string(),
+            transaction_validity_period: TRANSACTION_VALIDITY_PERIOD,
+            chunk_producer_kickout_threshold: CHUNK_PRODUCER_KICKOUT_THRESHOLD,
+            fishermen_threshold: FISHERMEN_THRESHOLD,
+            min_gas_price: MIN_GAS_PRICE,
+        };
+        config.init();
+        config
+    }
+
+    pub fn test(seeds: Vec<&str>, num_validator_seats: NumSeats) -> Self {
+        Self::test_with_seeds(seeds, num_validator_seats, vec![num_validator_seats])
+    }
+
+    pub fn test_free(seeds: Vec<&str>, num_validator_seats: NumSeats) -> Self {
+        let mut config =
+            Self::test_with_seeds(seeds, num_validator_seats, vec![num_validator_seats]);
+        config.runtime_config = RuntimeConfig::free();
+        config
+    }
+
+    pub fn test_sharded(
+        seeds: Vec<&str>,
+        num_validator_seats: NumSeats,
+        num_validator_seats_per_shard: Vec<NumSeats>,
+    ) -> Self {
+        Self::test_with_seeds(seeds, num_validator_seats, num_validator_seats_per_shard)
+    }
+}
+
 #[derive(Clone)]
 pub struct NearConfig {
     config: Config,
@@ -403,101 +485,6 @@ impl NearConfig {
     }
 }
 
-/// Account info for validators
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct AccountInfo {
-    pub account_id: AccountId,
-    pub public_key: PublicKey,
-    #[serde(with = "u128_dec_format")]
-    pub amount: Balance,
-}
-
-pub const CONFIG_VERSION: u32 = 1;
-
-/// Runtime configuration, defining genesis block.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct GenesisConfig {
-    /// This is a version of a genesis config structure this version of binary works with.
-    /// If the binary tries to load a JSON config with a different version it will panic.
-    /// It's not a major protocol version, but used for automatic config migrations using scripts.
-    pub config_version: u32,
-    /// Protocol version that this genesis works with.
-    pub protocol_version: u32,
-    /// Official time of blockchain start.
-    pub genesis_time: DateTime<Utc>,
-    /// ID of the blockchain. This must be unique for every blockchain.
-    /// If your testnet blockchains do not have unique chain IDs, you will have a bad time.
-    pub chain_id: String,
-    /// Number of block producer seats at genesis.
-    pub num_block_producer_seats: NumSeats,
-    /// Defines number of shards and number of block producer seats per each shard at genesis.
-    pub num_block_producer_seats_per_shard: Vec<NumSeats>,
-    /// Expected number of hidden validators per shard.
-    pub avg_hidden_validator_seats_per_shard: Vec<NumSeats>,
-    /// Enable dynamic re-sharding.
-    pub dynamic_resharding: bool,
-    /// Epoch length counted in block heights.
-    pub epoch_length: BlockHeightDelta,
-    /// Initial gas limit.
-    pub gas_limit: Gas,
-    /// Minimum gas price. It is also the initial gas price.
-    pub min_gas_price: Balance,
-    /// Criterion for kicking out block producers (this is a number between 0 and 100)
-    pub block_producer_kickout_threshold: u8,
-    /// Criterion for kicking out chunk producers (this is a number between 0 and 100)
-    pub chunk_producer_kickout_threshold: u8,
-    /// Gas price adjustment rate
-    pub gas_price_adjustment_rate: u8,
-    /// Runtime configuration (mostly economics constants).
-    pub runtime_config: RuntimeConfig,
-    /// List of initial validators.
-    pub validators: Vec<AccountInfo>,
-    /// Records in storage at genesis (get split into shards at genesis creation).
-    pub records: Vec<StateRecord>,
-    /// Number of blocks for which a given transaction is valid
-    pub transaction_validity_period: NumBlocks,
-    /// Developer reward percentage (this is a number between 0 and 100)
-    pub developer_reward_percentage: u8,
-    /// Protocol treasury percentage (this is a number between 0 and 100)
-    pub protocol_reward_percentage: u8,
-    /// Maximum inflation on the total supply every epoch (this is a number between 0 and 100)
-    pub max_inflation_rate: u8,
-    /// Total supply of tokens at genesis.
-    pub total_supply: u128,
-    /// Expected number of blocks per year
-    pub num_blocks_per_year: u64,
-    /// Protocol treasury account
-    pub protocol_treasury_account: AccountId,
-    /// Fishermen stake threshold.
-    #[serde(with = "u128_dec_format")]
-    pub fishermen_threshold: Balance,
-}
-
-pub fn get_initial_supply(records: &[StateRecord]) -> Balance {
-    let mut total_supply = 0;
-    for record in records {
-        if let StateRecord::Account { account, .. } = record {
-            total_supply += account.amount + account.locked;
-        }
-    }
-    total_supply
-}
-
-impl From<GenesisConfig> for ChainGenesis {
-    fn from(genesis_config: GenesisConfig) -> Self {
-        ChainGenesis::new(
-            genesis_config.genesis_time,
-            genesis_config.gas_limit,
-            genesis_config.min_gas_price,
-            genesis_config.total_supply,
-            genesis_config.max_inflation_rate,
-            genesis_config.gas_price_adjustment_rate,
-            genesis_config.transaction_validity_period,
-            genesis_config.epoch_length,
-        )
-    }
-}
-
 fn add_protocol_account(records: &mut Vec<StateRecord>) {
     let signer = InMemorySigner::from_seed(
         PROTOCOL_TREASURY_ACCOUNT,
@@ -511,133 +498,6 @@ fn add_protocol_account(records: &mut Vec<StateRecord>) {
         0,
         CryptoHash::default(),
     ));
-}
-
-const DEFAULT_TEST_CONTRACT: &'static [u8] =
-    include_bytes!("../../runtime/near-vm-runner/tests/res/test_contract_rs.wasm");
-
-impl GenesisConfig {
-    fn test_with_seeds(
-        seeds: Vec<&str>,
-        num_validator_seats: NumSeats,
-        num_validator_seats_per_shard: Vec<NumSeats>,
-    ) -> Self {
-        let mut validators = vec![];
-        let mut records = vec![];
-        let encoded_test_contract = to_base64(&DEFAULT_TEST_CONTRACT);
-        let code_hash = hash(&DEFAULT_TEST_CONTRACT);
-        for (i, &account) in seeds.iter().enumerate() {
-            let signer = InMemorySigner::from_seed(account, KeyType::ED25519, account);
-            let i = i as u64;
-            if i < num_validator_seats {
-                validators.push(AccountInfo {
-                    account_id: account.to_string(),
-                    public_key: signer.public_key.clone(),
-                    amount: TESTING_INIT_STAKE,
-                });
-            }
-            records.extend(
-                state_records_account_with_key(
-                    account,
-                    &signer.public_key.clone(),
-                    TESTING_INIT_BALANCE
-                        - if i < num_validator_seats { TESTING_INIT_STAKE } else { 0 },
-                    if i < num_validator_seats { TESTING_INIT_STAKE } else { 0 },
-                    code_hash,
-                )
-                .into_iter(),
-            );
-            records.push(StateRecord::Contract {
-                account_id: account.to_string(),
-                code: encoded_test_contract.clone(),
-            });
-        }
-        add_protocol_account(&mut records);
-        let total_supply = get_initial_supply(&records);
-        GenesisConfig {
-            protocol_version: PROTOCOL_VERSION,
-            config_version: CONFIG_VERSION,
-            genesis_time: Utc::now(),
-            chain_id: random_chain_id(),
-            num_block_producer_seats: num_validator_seats,
-            num_block_producer_seats_per_shard: num_validator_seats_per_shard.clone(),
-            avg_hidden_validator_seats_per_shard: num_validator_seats_per_shard
-                .iter()
-                .map(|_| 0)
-                .collect(),
-            dynamic_resharding: false,
-            epoch_length: FAST_EPOCH_LENGTH,
-            gas_limit: INITIAL_GAS_LIMIT,
-            gas_price_adjustment_rate: GAS_PRICE_ADJUSTMENT_RATE,
-            block_producer_kickout_threshold: BLOCK_PRODUCER_KICKOUT_THRESHOLD,
-            runtime_config: Default::default(),
-            validators,
-            records,
-            developer_reward_percentage: DEVELOPER_PERCENT,
-            protocol_reward_percentage: PROTOCOL_PERCENT,
-            max_inflation_rate: MAX_INFLATION_RATE,
-            total_supply,
-            num_blocks_per_year: NUM_BLOCKS_PER_YEAR,
-            protocol_treasury_account: PROTOCOL_TREASURY_ACCOUNT.to_string(),
-            transaction_validity_period: TRANSACTION_VALIDITY_PERIOD,
-            chunk_producer_kickout_threshold: CHUNK_PRODUCER_KICKOUT_THRESHOLD,
-            fishermen_threshold: FISHERMEN_THRESHOLD,
-            min_gas_price: MIN_GAS_PRICE,
-        }
-    }
-
-    pub fn test(seeds: Vec<&str>, num_validator_seats: NumSeats) -> Self {
-        Self::test_with_seeds(seeds, num_validator_seats, vec![num_validator_seats])
-    }
-
-    pub fn test_free(seeds: Vec<&str>, num_validator_seats: NumSeats) -> Self {
-        let mut config =
-            Self::test_with_seeds(seeds, num_validator_seats, vec![num_validator_seats]);
-        config.runtime_config = RuntimeConfig::free();
-        config
-    }
-
-    pub fn test_sharded(
-        seeds: Vec<&str>,
-        num_validator_seats: NumSeats,
-        num_validator_seats_per_shard: Vec<NumSeats>,
-    ) -> Self {
-        Self::test_with_seeds(seeds, num_validator_seats, num_validator_seats_per_shard)
-    }
-
-    /// Reads GenesisConfig from a file.
-    pub fn from_file(path: &PathBuf) -> Self {
-        let mut file = File::open(path).expect("Could not open genesis config file.");
-        let mut content = String::new();
-        file.read_to_string(&mut content).expect("Could not read from genesis config file.");
-        GenesisConfig::from(content.as_str())
-    }
-
-    /// Writes GenesisConfig to the file.
-    pub fn write_to_file(&self, path: &Path) {
-        let mut file = File::create(path).expect("Failed to create / write a genesis config file.");
-        let str =
-            serde_json::to_string_pretty(self).expect("Error serializing the genesis config.");
-        if let Err(err) = file.write_all(str.as_bytes()) {
-            panic!("Failed to write a genesis config file {}", err);
-        }
-    }
-}
-
-impl From<&str> for GenesisConfig {
-    fn from(config: &str) -> Self {
-        let mut config: GenesisConfig =
-            serde_json::from_str(config).expect("Failed to deserialize the genesis config.");
-        if config.protocol_version != PROTOCOL_VERSION {
-            panic!(format!(
-                "Incorrect version of genesis config {} expected {}",
-                config.protocol_version, PROTOCOL_VERSION
-            ));
-        }
-        let total_supply = get_initial_supply(&config.records);
-        config.total_supply = total_supply;
-        config
-    }
 }
 
 fn random_chain_id() -> String {
@@ -763,11 +623,10 @@ pub fn init_configs(
                 CryptoHash::default(),
             );
             add_protocol_account(&mut records);
-            let total_supply = get_initial_supply(&records);
 
-            let genesis_config = GenesisConfig {
+            let mut genesis_config = GenesisConfig {
                 protocol_version: PROTOCOL_VERSION,
-                config_version: CONFIG_VERSION,
+                config_version: GENESIS_CONFIG_VERSION,
                 genesis_time: Utc::now(),
                 chain_id,
                 num_block_producer_seats: NUM_BLOCK_PRODUCER_SEATS,
@@ -792,13 +651,14 @@ pub fn init_configs(
                 developer_reward_percentage: DEVELOPER_PERCENT,
                 protocol_reward_percentage: PROTOCOL_PERCENT,
                 max_inflation_rate: MAX_INFLATION_RATE,
-                total_supply,
+                total_supply: 0,
                 num_blocks_per_year: NUM_BLOCKS_PER_YEAR,
                 protocol_treasury_account: account_id,
                 chunk_producer_kickout_threshold: CHUNK_PRODUCER_KICKOUT_THRESHOLD,
                 fishermen_threshold: FISHERMEN_THRESHOLD,
                 min_gas_price: MIN_GAS_PRICE,
             };
+            genesis_config.init();
             genesis_config.write_to_file(&dir.join(config.genesis_file));
             info!(target: "near", "Generated node key, validator key, genesis file in {}", dir.to_str().unwrap());
         }
@@ -843,7 +703,7 @@ pub fn create_testnet_configs_from_seeds(
             config.network.skip_sync_wait = num_validator_seats == 1;
         }
         config.consensus.min_num_peers =
-            cmp::min(num_validator_seats as usize - 1, config.consensus.min_num_peers);
+            std::cmp::min(num_validator_seats as usize - 1, config.consensus.min_num_peers);
         configs.push(config);
     }
     (configs, validator_signers, network_signers, genesis_config)
@@ -940,7 +800,7 @@ mod test {
     fn test_deserialize_state() {
         let genesis_config = testnet_genesis();
         assert_eq!(genesis_config.protocol_version, PROTOCOL_VERSION);
-        assert_eq!(genesis_config.config_version, CONFIG_VERSION);
+        assert_eq!(genesis_config.config_version, GENESIS_CONFIG_VERSION);
         assert!(genesis_config.total_supply > 0);
     }
 }
