+++ test-utils/testlib/Cargo.toml
@@ -22,6 +22,7 @@ assert_matches = "1.3.0"
 
 borsh = "0.2.10"
 
+near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
