+++ test-utils/testlib/src/node/mod.rs
@@ -5,17 +5,18 @@ use std::sync::Arc;
 use std::sync::RwLock;
 
 use near::config::{
-    create_testnet_configs, create_testnet_configs_from_seeds, Config, GenesisConfig,
+    create_testnet_configs, create_testnet_configs_from_seeds, Config, GenesisConfigExt,
 };
 use near::NearConfig;
+use near_chain_configs::GenesisConfig;
 use near_crypto::{InMemorySigner, Signer};
 use near_jsonrpc::ServerError;
 use near_primitives::serialize::to_base64;
+use near_primitives::state_record::StateRecord;
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::{AccountId, Balance, NumSeats};
 use near_primitives::validator_signer::InMemoryValidatorSigner;
 use near_primitives::views::AccountView;
-use node_runtime::StateRecord;
 
 pub use crate::node::process_node::ProcessNode;
 pub use crate::node::runtime_node::RuntimeNode;
